Correct types for enumeration types and constant values Change the constant values for the protocol in the IPv4 packet to be in an enumeration type

+++ trunk/PacketCaptureAnalyser/Ethernet Frame/IPv4 Packet/ICMPv4 Packet/ICMPv4PacketConstants.cs
@@ -33,6 +33,6 @@ namespace EthernetFrameNamespace.IPv4PacketNamespace.ICMPv4PacketNamespace
 
         //Length
 
-        public const int ICMPv4PacketHeaderLength = 4;
+        public const ushort ICMPv4PacketHeaderLength = 4;
     }
 }
